WL-3119 Add scope to define a METHOD for an event
To take advantage of some of the parts of the iTIP spec this extends the calendaring apis
to define a method on an iCal event.

This introduces certain requirements on the events
for example you cannot define REQUEST method for an event with no participants.

Similarly an event with the CANCEL method must have the SEQUENCE property defined.

Also adds a mechanism in the API to add participants with the chair role.