WL-3119 Specify calendar method
Extend the constructor for a Calendar to include the ITIP  method.

In particular this allows to specify special methods for ical attachments
such as `REQUEST` or `CANCEL` to improve the interaction with mail clients.

N.B. `REQUEST` method requires events have attendees and `CANCEL` requires
events have a `SEQUENCE` property.